+++ RELEASES.md
@@ -15,6 +15,7 @@ Bug fixes
 
 * Constants are no longer propagated by the optimizer if shadowed by a non-constant variable.
 * Constants passed as the `this` parameter to Rhai functions now throws an error if assigned to.
+* Generic type parameter of `Engine::register_iterator` is `IntoIterator` instead of `Iterator`.
 
 Breaking changes
 ----------------